+++ src/errors/index.ts
@@ -1,0 +1,5 @@
+import { catchController } from './catchController.js';
+import { SERVER_ERROR, INVALID_URL, USER_NOT_EXIST, USER_ID_NOT_VALID, INVALID_DATA } from './constants.js';
+import { CustomError } from './customError.js';
+
+export { catchController, SERVER_ERROR, INVALID_URL, USER_NOT_EXIST, USER_ID_NOT_VALID, INVALID_DATA, CustomError };
